Make sure archive ends up at the right place
We achieve this by narrowing archive buffer before archiving to node we
want our archived entry to and up at. I was not sure this is going to
work, but it does.

This was the one major limitation of original design, so happy I ended
up fixing it with a rather simple change.